build: split up plugin logic
