--- buildSrc/src/main/kotlin/io/embrace/internal/Versions.kt
@@ -1,11 +1,0 @@
-package io.embrace.internal
-
-/**
- * Defines dependency versions that are used in the project that cannot be defined in libs.versions.toml
- */
-object Versions {
-    const val COMPILE_SDK = 35
-    const val MIN_SDK = 21
-    const val MIN_COMPILE_SDK = 34
-    const val NDK = "22.1.7171670"
-}
